+++ module/helpers/templates.mjs
@@ -39,6 +39,7 @@ export const preloadHandlebarsTemplates = async function () {
     "systems/mist-engine-fvtt/templates/actor/parts/npc-challenges-edit-partial.hbs",
     "systems/mist-engine-fvtt/templates/dice-roll-app/might.hbs",
     "systems/mist-engine-fvtt/templates/dice-roll-app/gm-roll-confirm.hbs",
-    "systems/mist-engine-fvtt/templates/overlay/tag-entry.hbs"
+    "systems/mist-engine-fvtt/templates/overlay/tag-entry.hbs",
+    "systems/mist-engine-fvtt/templates/overlay/scene-tags-overlay.hbs"
   ]);
 };